refactor: weatherInfoData移動
